requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.7
celery>=5.4
redis>=5.0
pytest>=8.2
pytest-django>=4.8
black>=24.8
isort>=5.13
flake8>=7.1
